Used more convinient uintX types for numbers in contracts

+++ src/deploy/deploy-home/contracts/Bridge.sol
@@ -4,16 +4,16 @@ import './openzeppelin-solidity/contracts/token/ERC20/IERC20.sol';
 import "./MessageHandler.sol";
 
 contract Bridge is MessageHandler {
-    event ExchangeRequest(uint value, uint nonce);
+    event ExchangeRequest(uint96 value, uint32 nonce);
 
     mapping(bytes32 => bool) usedExchangeRanges;
 
     constructor(
-        uint threshold,
+        uint16 threshold,
         address[] memory validators,
         address _tokenContract,
-        uint[2] memory limits,
-        uint rangeSize,
+        uint96[2] memory limits,
+        uint16 rangeSize,
         bool closeEpoch
     ) public {
         require(validators.length > 0);
@@ -43,10 +43,10 @@ contract Bridge is MessageHandler {
         emit NewEpoch(0, 1);
     }
 
-    function exchange(uint value) public ready {
+    function exchange(uint96 value) public ready {
         require(value >= minTxLimit && value >= 10 ** 10 && value <= maxTxLimit);
 
-        uint txRange = (block.number - getStartBlock()) / getRangeSize();
+        uint32 txRange = (uint32(block.number) - getStartBlock()) / uint32(getRangeSize());
         if (!usedExchangeRanges[keccak256(abi.encodePacked(txRange, epoch))]) {
             usedExchangeRanges[keccak256(abi.encodePacked(txRange, epoch))] = true;
             states[epoch].nonce++;
